+++ src/VersaORM.php
@@ -87,7 +87,7 @@ use function strlen;
  * NOTA: Todos los métodos de consulta y manipulación de datos
  * están ahora en VersaModel para una arquitectura más limpia.
  *
- * @version 1.6.0
+ * @version 1.7.0
  *
  * @author  VersaORM Team
  * @license MIT